disp: msm: sde: fix ich reset override logic
This change fixes the detection logic for overriding the ich
reset in single dsc case. In the previous logic ich reset override
was not getting triggered when partial update on single dsc was
enabled. This override is required to change the
default Hw behavior of changing the ich reset position.

For single dsc case partial update case when DSC encoder
configuration changes from 2 slices to single slice
this override is required.

Change-Id: I435dc7ff10c9fb0edb8e40e6701608aa22136981
Signed-off-by: Abhijit Kulkarni <kabhijit@codeaurora.org>
Signed-off-by: Jayaprakash <jmadiset@codeaurora.org>